0606. 13:28
권형님 파일 정리 및 지정 appColor 추가
Color 쓰는 법 : color : AppColors.지정색
