+++ App/pick_caffeine_app/lib/app_colors.dart
@@ -1,0 +1,13 @@
+import 'dart:ui';
+
+class AppColors {
+  AppColors._();
+
+  static const Color black = Color(0xff000000);
+  static const Color white = Color(0xffFAFAFA);
+  static const Color brown = Color(0xff6D4D41);
+  static const Color grey = Color(0xffBDB6B6);
+  static const Color lightbrown = Color(0xffD7A86E);
+  static const Color red = Color(0xffE12424);
+  static const Color lightpick = Color(0xffFFF0F0);
+}
