+++ App/pick_caffeine_app/lib/view/admin/admin_inquiry_list.dart
@@ -16,9 +16,9 @@
 
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:flutter_slidable/flutter_slidable.dart'; 
-import 'package:pick_caffein/vm/getx_controller.dart'; 
-import 'package:pick_caffein/view/widgets/admin_bottom_tab.dart'; 
+import 'package:flutter_slidable/flutter_slidable.dart';
+
+import 'package:pick_caffeine_app/kwonhyoung_controller.dart';
 
 class InquiryReport extends StatelessWidget {
   InquiryReport({super.key});
@@ -33,16 +33,18 @@ class InquiryReport extends StatelessWidget {
       backgroundColor: Colors.grey[100],
       body: Column(
         children: [
-          _buildTopImageWithText(),     // 상단 이미지 + 제목 텍스트
-          _buildTabBar(),               // 탭 버튼 (문의내역 / 답변내역)
-          Expanded(                     // 실제 콘텐츠 영역
-            child: Obx(() =>
-              tabIndex.value == 0
-                ? _buildInquiryList()   // 문의내역 탭일 때
-                : _buildAnswerList()    // 답변내역 탭일 때
+          _buildTopImageWithText(), // 상단 이미지 + 제목 텍스트
+          _buildTabBar(), // 탭 버튼 (문의내역 / 답변내역)
+          Expanded(
+            // 실제 콘텐츠 영역
+            child: Obx(
+              () =>
+                  tabIndex.value == 0
+                      ? _buildInquiryList() // 문의내역 탭일 때
+                      : _buildAnswerList(), // 답변내역 탭일 때
             ),
           ),
-          BottomTabbar(selectedIndex: 1), // 하단 네비게이션 탭바 (문의 관리 강조)
+          // Bottom(selectedIndex: 1), // 하단 네비게이션 탭바 (문의 관리 강조)
         ],
       ),
     );
@@ -60,53 +62,63 @@ class InquiryReport extends StatelessWidget {
         ),
         Container(
           padding: EdgeInsets.symmetric(vertical: 12),
-          child: Obx(() => Text(
-            tabIndex.value == 0 ? '문의내역' : '답변내역',
-            style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold, color: Color(0xFF8B4513)),
-          )),
-        )
+          child: Obx(
+            () => Text(
+              tabIndex.value == 0 ? '문의내역' : '답변내역',
+              style: TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.bold,
+                color: Color(0xFF8B4513),
+              ),
+            ),
+          ),
+        ),
       ],
     );
   }
 
   // 탭 버튼 영역 (문의내역 / 답변내역)
   Widget _buildTabBar() {
-    return Obx(() => Row(
-      children: [
-        // 문의내역 탭 버튼
-        Expanded(
-          child: TextButton(
-            onPressed: () => tabIndex.value = 0,
-            style: TextButton.styleFrom(
-              backgroundColor: tabIndex.value == 0 ? Color(0xFF8B4513) : Colors.white,
-            ),
-            child: Text(
-              '문의내역',
-              style: TextStyle(
-                color: tabIndex.value == 0 ? Colors.white : Colors.grey[700],
-                fontWeight: FontWeight.w600,
+    return Obx(
+      () => Row(
+        children: [
+          // 문의내역 탭 버튼
+          Expanded(
+            child: TextButton(
+              onPressed: () => tabIndex.value = 0,
+              style: TextButton.styleFrom(
+                backgroundColor:
+                    tabIndex.value == 0 ? Color(0xFF8B4513) : Colors.white,
+              ),
+              child: Text(
+                '문의내역',
+                style: TextStyle(
+                  color: tabIndex.value == 0 ? Colors.white : Colors.grey[700],
+                  fontWeight: FontWeight.w600,
+                ),
               ),
             ),
           ),
-        ),
-        // 답변내역 탭 버튼
-        Expanded(
-          child: TextButton(
-            onPressed: () => tabIndex.value = 1,
-            style: TextButton.styleFrom(
-              backgroundColor: tabIndex.value == 1 ? Color(0xFF8B4513) : Colors.white,
-            ),
-            child: Text(
-              '답변내역',
-              style: TextStyle(
-                color: tabIndex.value == 1 ? Colors.white : Colors.grey[700],
-                fontWeight: FontWeight.w600,
+          // 답변내역 탭 버튼
+          Expanded(
+            child: TextButton(
+              onPressed: () => tabIndex.value = 1,
+              style: TextButton.styleFrom(
+                backgroundColor:
+                    tabIndex.value == 1 ? Color(0xFF8B4513) : Colors.white,
+              ),
+              child: Text(
+                '답변내역',
+                style: TextStyle(
+                  color: tabIndex.value == 1 ? Colors.white : Colors.grey[700],
+                  fontWeight: FontWeight.w600,
+                ),
               ),
             ),
           ),
-        ),
-      ],
-    ));
+        ],
+      ),
+    );
   }
 
   // 문의 내역 리스트 렌더링 (답변 처리 + 반려 처리 포함)
@@ -134,7 +146,8 @@ class InquiryReport extends StatelessWidget {
               children: [
                 // 슬라이드 시 반려 버튼 노출
                 SlidableAction(
-                  onPressed: (_) => controller.deleteInquiry(inquiry.inquiryNum),
+                  onPressed:
+                      (_) => controller.deleteInquiry(inquiry.inquiryNum),
                   backgroundColor: Colors.red[300]!,
                   foregroundColor: Colors.white,
                   icon: Icons.close,
@@ -149,14 +162,20 @@ class InquiryReport extends StatelessWidget {
                 borderRadius: BorderRadius.circular(12),
               ),
               child: ListTile(
-                contentPadding: EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+                contentPadding: EdgeInsets.symmetric(
+                  horizontal: 16,
+                  vertical: 10,
+                ),
                 title: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     // 문의 내용
                     Text(
                       inquiry.inquiryContent,
-                      style: TextStyle(fontSize: 16, fontWeight: FontWeight.w600),
+                      style: TextStyle(
+                        fontSize: 16,
+                        fontWeight: FontWeight.w600,
+                      ),
                     ),
                     SizedBox(height: 4),
                     // ID + 작성일 표시
@@ -189,59 +208,70 @@ class InquiryReport extends StatelessWidget {
 
   // 답변 내역 리스트 (answerContent가 null이 아닌 항목만 표시)
   Widget _buildAnswerList() {
-  return Obx(() {
-    final answers = controller.inquiryList.where((e) => e.response != null).toList();
+    return Obx(() {
+      final answers =
+          controller.inquiryList.where((e) => e.response != null).toList();
 
-    if (answers.isEmpty) {
-      return Center(child: Text('답변 완료된 내역이 없습니다.'));
-    }
+      if (answers.isEmpty) {
+        return Center(child: Text('답변 완료된 내역이 없습니다.'));
+      }
 
-    return ListView.builder(
-      padding: EdgeInsets.all(16),
-      itemCount: answers.length,
-      itemBuilder: (context, index) {
-        final inquiry = answers[index];
+      return ListView.builder(
+        padding: EdgeInsets.all(16),
+        itemCount: answers.length,
+        itemBuilder: (context, index) {
+          final inquiry = answers[index];
 
-        return Container(
-          margin: EdgeInsets.only(bottom: 12),
-          decoration: BoxDecoration(
-            color: Color(0xFFFFF1EC),
-            borderRadius: BorderRadius.circular(12),
-            boxShadow: [
-              BoxShadow(color: Colors.black12, blurRadius: 4, offset: Offset(0, 2)),
-            ],
-          ),
-          child: ListTile(
-            contentPadding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-            title: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                // ID, 닉네임, 답변일자 표시
-                Text(
-                  '등록일: ${_formatDate((inquiry.response is DateTime ? inquiry.response as DateTime : DateTime.now()))}',
-                  style: TextStyle(fontSize: 12, color: Colors.grey[600]),
+          return Container(
+            margin: EdgeInsets.only(bottom: 12),
+            decoration: BoxDecoration(
+              color: Color(0xFFFFF1EC),
+              borderRadius: BorderRadius.circular(12),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.black12,
+                  blurRadius: 4,
+                  offset: Offset(0, 2),
                 ),
-                Text(
-                  'ID/닉네임: ${inquiry.userId} / ${inquiry.userNickname}',
-                  style: TextStyle(fontSize: 13, fontWeight: FontWeight.w500),
-                ),
-                SizedBox(height: 8),
-
-                // 문의 및 답변 내용
-                Text('문의: ${inquiry.inquiryContent}',
-                    style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600)),
-                SizedBox(height: 4),
-                Text('답변: ${inquiry.response}',
-                    style: TextStyle(fontSize: 14, color: Colors.black87)),
               ],
             ),
-          ),
-        );
-      },
-    );
-  });
-}
+            child: ListTile(
+              contentPadding: EdgeInsets.symmetric(
+                horizontal: 16,
+                vertical: 12,
+              ),
+              title: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  // ID, 닉네임, 답변일자 표시
+                  Text(
+                    '등록일: ${_formatDate((inquiry.response is DateTime ? inquiry.response as DateTime : DateTime.now()))}',
+                    style: TextStyle(fontSize: 12, color: Colors.grey[600]),
+                  ),
+                  Text(
+                    'ID/닉네임: ${inquiry.userId} / ${inquiry.userNickname}',
+                    style: TextStyle(fontSize: 13, fontWeight: FontWeight.w500),
+                  ),
+                  SizedBox(height: 8),
 
+                  // 문의 및 답변 내용
+                  Text(
+                    '문의: ${inquiry.inquiryContent}',
+                    style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
+                  ),
+                  SizedBox(height: 4),
+                  Text(
+                    '답변: ${inquiry.response}',
+                    style: TextStyle(fontSize: 14, color: Colors.black87),
+                  ),
+                ],
+              ),
+            ),
+          );
+        },
+      );
+    });
+  }
 
   // 답변 처리 다이얼로그 팝업
   void _showAnswerDialog(int inquiryNum) {
