+++ App/pick_caffeine_app/lib/view/admin/admin_report_list.dart
@@ -18,9 +18,10 @@
 
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:pick_caffein/model/declaration_model.dart';
-import 'package:pick_caffein/view/inquiry_report.dart';
-import 'package:pick_caffein/vm/getx_controller.dart';
+import 'package:pick_caffeine_app/admin_model.dart';
+
+import 'package:pick_caffeine_app/kwonhyoung_controller.dart';
+import 'package:pick_caffeine_app/view/admin/admin_inquiry_list.dart';
 
 // 관리자 첫페이지(유저 신고관리)
 class AdminReportScreen extends StatelessWidget {
@@ -50,9 +51,7 @@ class AdminReportScreen extends StatelessWidget {
         SizedBox(
           width: double.infinity,
           height: 150,
-          child: Image.asset('images/cafe.png',
-            fit: BoxFit.cover,
-          ),
+          child: Image.asset('images/cafe.png', fit: BoxFit.cover),
         ),
       ],
     );
@@ -66,20 +65,18 @@ class AdminReportScreen extends StatelessWidget {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
-          Obx(() => Text(
-            '유저 수: ${controller.userCount.value}명',
-            style: TextStyle(
-              fontSize: 14,
-              fontWeight: FontWeight.w600,
+          Obx(
+            () => Text(
+              '유저 수: ${controller.userCount.value}명',
+              style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
             ),
-          )),
-          Obx(() => Text(
-            '매장 수: ${controller.storeCount.value}개',
-            style: TextStyle(
-              fontSize: 14,
-              fontWeight: FontWeight.w600,
+          ),
+          Obx(
+            () => Text(
+              '매장 수: ${controller.storeCount.value}개',
+              style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
             ),
-          )),
+          ),
         ],
       ),
     );
@@ -95,11 +92,7 @@ class AdminReportScreen extends StatelessWidget {
         unselectedLabelColor: Colors.grey[600],
         indicatorColor: Color(0xFF8B4513),
         indicatorWeight: 3,
-        tabs: [
-          Tab(text: "제재 등록"),
-          Tab(text: "신고접수 관리"),
-          Tab(text: "제재 유저 목록"),
-        ],
+        tabs: [Tab(text: "제재 등록"), Tab(text: "신고접수 관리"), Tab(text: "제재 유저 목록")],
       ),
     );
   }
@@ -110,9 +103,9 @@ class AdminReportScreen extends StatelessWidget {
       child: TabBarView(
         controller: controller.tabController,
         children: [
-          _buildSanctionRegistrationTab(),  // 제재 등록
-          _buildReportManagementTab(),      // 신고접수 관리
-          _buildSanctionedUsersTab(),       // 제재 유저 목록
+          _buildSanctionRegistrationTab(), // 제재 등록
+          _buildReportManagementTab(), // 신고접수 관리
+          _buildSanctionedUsersTab(), // 제재 유저 목록
         ],
       ),
     );
@@ -126,7 +119,11 @@ class AdminReportScreen extends StatelessWidget {
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              Icon(Icons.warning_amber_outlined, size: 80, color: Colors.grey[400]),
+              Icon(
+                Icons.warning_amber_outlined,
+                size: 80,
+                color: Colors.grey[400],
+              ),
               SizedBox(height: 16),
               Text(
                 '제재 처리할 신고를 선택해주세요',
@@ -141,7 +138,10 @@ class AdminReportScreen extends StatelessWidget {
                     borderRadius: BorderRadius.circular(8),
                   ),
                 ),
-                child: Text('신고 목록으로 이동', style: TextStyle(color: Colors.white)),
+                child: Text(
+                  '신고 목록으로 이동',
+                  style: TextStyle(color: Colors.white),
+                ),
               ),
             ],
           ),
@@ -149,7 +149,7 @@ class AdminReportScreen extends StatelessWidget {
       }
 
       final declaration = controller.selectedDeclaration.value!;
-      
+
       return SingleChildScrollView(
         padding: EdgeInsets.all(16),
         child: Column(
@@ -158,11 +158,11 @@ class AdminReportScreen extends StatelessWidget {
             // 신고 정보 카드
             _buildSanctionInfoCard(declaration),
             SizedBox(height: 20),
-            
+
             // 제재 옵션 선택
             _buildSanctionOptionsCard(),
             SizedBox(height: 20),
-            
+
             // 제재 등록 버튼
             _buildSanctionActionButtons(),
           ],
@@ -195,7 +195,9 @@ class AdminReportScreen extends StatelessWidget {
                 Container(
                   padding: EdgeInsets.symmetric(horizontal: 12, vertical: 6),
                   decoration: BoxDecoration(
-                    color: controller.getStatusColor(declaration.declarationState),
+                    color: controller.getStatusColor(
+                      declaration.declarationState,
+                    ),
                     borderRadius: BorderRadius.circular(20),
                   ),
                   child: Text(
@@ -210,7 +212,7 @@ class AdminReportScreen extends StatelessWidget {
               ],
             ),
             SizedBox(height: 16),
-            
+
             // 날짜 정보
             Row(
               children: [
@@ -223,7 +225,7 @@ class AdminReportScreen extends StatelessWidget {
               ],
             ),
             SizedBox(height: 12),
-            
+
             // 신고 내용
             Container(
               width: double.infinity,
@@ -281,7 +283,7 @@ class AdminReportScreen extends StatelessWidget {
               ),
             ),
             SizedBox(height: 20),
-            
+
             // 제재 구분 드롭다운
             Column(
               crossAxisAlignment: CrossAxisAlignment.start,
@@ -302,27 +304,32 @@ class AdminReportScreen extends StatelessWidget {
                     borderRadius: BorderRadius.circular(8),
                     color: Colors.white,
                   ),
-                  child: Obx(() => DropdownButton<String>(
-                    value: controller.selectedSanctionType.value,
-                    isExpanded: true,
-                    underline: SizedBox(),
-                    items: ['1차 제재', '2차 제재']
-                        .map((type) => DropdownMenuItem(
-                              value: type,
-                              child: Text(type),
-                            ))
-                        .toList(),
-                    onChanged: (value) {
-                      if (value != null) {
-                        controller.setSanctionType(value);
-                      }
-                    },
-                  )),
+                  child: Obx(
+                    () => DropdownButton<String>(
+                      value: controller.selectedSanctionType.value,
+                      isExpanded: true,
+                      underline: SizedBox(),
+                      items:
+                          ['1차 제재', '2차 제재']
+                              .map(
+                                (type) => DropdownMenuItem(
+                                  value: type,
+                                  child: Text(type),
+                                ),
+                              )
+                              .toList(),
+                      onChanged: (value) {
+                        if (value != null) {
+                          controller.setSanctionType(value);
+                        }
+                      },
+                    ),
+                  ),
                 ),
               ],
             ),
             SizedBox(height: 16),
-            
+
             // 제재 기간 드롭다운
             Container(
               padding: EdgeInsets.symmetric(horizontal: 16),
@@ -331,25 +338,30 @@ class AdminReportScreen extends StatelessWidget {
                 borderRadius: BorderRadius.circular(8),
                 color: Colors.white,
               ),
-              child: Obx(() => DropdownButton<String>(
-                value: controller.selectedSanctionPeriod.value,
-                isExpanded: true,
-                underline: SizedBox(),
-                items: ['1일', '3일', '7일', '30일', '영구정지']
-                    .map((period) => DropdownMenuItem(
-                          value: period,
-                          child: Text(period),
-                        ))
-                    .toList(),
-                onChanged: (value) {
-                  if (value != null) {
-                    controller.setSanctionPeriod(value);
-                  }
-                },
-              )),
+              child: Obx(
+                () => DropdownButton<String>(
+                  value: controller.selectedSanctionPeriod.value,
+                  isExpanded: true,
+                  underline: SizedBox(),
+                  items:
+                      ['1일', '3일', '7일', '30일', '영구정지']
+                          .map(
+                            (period) => DropdownMenuItem(
+                              value: period,
+                              child: Text(period),
+                            ),
+                          )
+                          .toList(),
+                  onChanged: (value) {
+                    if (value != null) {
+                      controller.setSanctionPeriod(value);
+                    }
+                  },
+                ),
+              ),
             ),
             SizedBox(height: 20),
-            
+
             // 제재 적용일
             Container(
               padding: EdgeInsets.all(16),
@@ -363,20 +375,19 @@ class AdminReportScreen extends StatelessWidget {
                 children: [
                   Text(
                     '제재 적용일: ${_formatDate(DateTime.now())}',
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: Colors.grey[700],
-                    ),
+                    style: TextStyle(fontSize: 14, color: Colors.grey[700]),
                   ),
                   SizedBox(height: 8),
-                  Obx(() => Text(
-                    '선택된 제재: ${controller.generateSanctionContent()}',
-                    style: TextStyle(
-                      fontSize: 14,
-                      fontWeight: FontWeight.w600,
-                      color: Colors.red[700],
+                  Obx(
+                    () => Text(
+                      '선택된 제재: ${controller.generateSanctionContent()}',
+                      style: TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                        color: Colors.red[700],
+                      ),
                     ),
-                  )),
+                  ),
                 ],
               ),
             ),
@@ -419,7 +430,8 @@ class AdminReportScreen extends StatelessWidget {
             controller.updateDeclaration(
               reviewNum: declaration.reviewNum,
               userId: declaration.userId,
-              declarationDate: declaration.declarationDate.toIso8601String().split('T')[0],
+              declarationDate:
+                  declaration.declarationDate.toIso8601String().split('T')[0],
               declarationContent: declaration.declarationContent,
               declarationState: '처리완료',
               sanctionContent: controller.generateSanctionContent(),
@@ -461,16 +473,21 @@ class AdminReportScreen extends StatelessWidget {
               );
             }
 
-            final pendingDeclarations = controller.declarations
-                .where((d) => d.sanctionContent == null)
-                .toList();
+            final pendingDeclarations =
+                controller.declarations
+                    .where((d) => d.sanctionContent == null)
+                    .toList();
 
             if (pendingDeclarations.isEmpty) {
               return Center(
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
-                    Icon(Icons.check_circle_outline, size: 80, color: Colors.green[400]),
+                    Icon(
+                      Icons.check_circle_outline,
+                      size: 80,
+                      color: Colors.green[400],
+                    ),
                     SizedBox(height: 16),
                     Text(
                       '처리 대기중인 신고가 없습니다.',
@@ -554,7 +571,9 @@ class AdminReportScreen extends StatelessWidget {
                   Container(
                     padding: EdgeInsets.symmetric(horizontal: 10, vertical: 4),
                     decoration: BoxDecoration(
-                      color: controller.getStatusColor(declaration.declarationState),
+                      color: controller.getStatusColor(
+                        declaration.declarationState,
+                      ),
                       borderRadius: BorderRadius.circular(12),
                     ),
                     child: Text(
@@ -571,14 +590,22 @@ class AdminReportScreen extends StatelessWidget {
               SizedBox(height: 12),
               Row(
                 children: [
-                  Icon(Icons.calendar_today_outlined, size: 14, color: Colors.grey[600]),
+                  Icon(
+                    Icons.calendar_today_outlined,
+                    size: 14,
+                    color: Colors.grey[600],
+                  ),
                   SizedBox(width: 4),
                   Text(
                     'Date: ${_formatDate(declaration.declarationDate)}',
                     style: TextStyle(fontSize: 12, color: Colors.grey[600]),
                   ),
                   SizedBox(width: 16),
-                  Icon(Icons.report_outlined, size: 14, color: Colors.grey[600]),
+                  Icon(
+                    Icons.report_outlined,
+                    size: 14,
+                    color: Colors.grey[600],
+                  ),
                   SizedBox(width: 4),
                   Text(
                     '리뷰번호: ${declaration.reviewNum}',
@@ -626,80 +653,91 @@ class AdminReportScreen extends StatelessWidget {
     );
   }
 
-// 제재 유저 목록 탭
-Widget _buildSanctionedUsersTab() {
-  return Column(
-    children: [
-      // 상단 정보 바
-      Container(
-        padding: EdgeInsets.all(16),
-        color: Colors.white,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Obx(() => Text(
-              '제재 유저 수: ${controller.sanctionedUserCount.value}명',
-              style: TextStyle(
-                fontSize: 14,
-                fontWeight: FontWeight.w600,
-                color: Colors.red[700],
-              ),
-            )),
-            IconButton(
-              onPressed: controller.refreshData,
-              icon: Icon(Icons.refresh, color: Color(0xFF8B4513)),
-            ),
-          ],
-        ),
-      ),
-      
-      // 제재 유저 리스트
-      Expanded(
-        child: Obx(() {
-          if (controller.isLoading.value) {
-            return Center(
-              child: CircularProgressIndicator(color: Color(0xFF8B4513)),
-            );
-          }
-
-          // 제재된 유저들 (sanctionContent가 있는 경우)
-          final sanctionedDeclarations = controller.declarations
-              .where((d) => d.sanctionContent != null && d.sanctionContent!.isNotEmpty)
-              .toList();
-
-          if (sanctionedDeclarations.isEmpty) {
-            return Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Icon(Icons.person_off_outlined, size: 80, color: Colors.grey[400]),
-                  SizedBox(height: 16),
-                  Text(
-                    '제재중인 유저가 없습니다.',
-                    style: TextStyle(fontSize: 16, color: Colors.grey[600]),
+  // 제재 유저 목록 탭
+  Widget _buildSanctionedUsersTab() {
+    return Column(
+      children: [
+        // 상단 정보 바
+        Container(
+          padding: EdgeInsets.all(16),
+          color: Colors.white,
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Obx(
+                () => Text(
+                  '제재 유저 수: ${controller.sanctionedUserCount.value}명',
+                  style: TextStyle(
+                    fontSize: 14,
+                    fontWeight: FontWeight.w600,
+                    color: Colors.red[700],
                   ),
-                ],
+                ),
+              ),
+              IconButton(
+                onPressed: controller.refreshData,
+                icon: Icon(Icons.refresh, color: Color(0xFF8B4513)),
+              ),
+            ],
+          ),
+        ),
+
+        // 제재 유저 리스트
+        Expanded(
+          child: Obx(() {
+            if (controller.isLoading.value) {
+              return Center(
+                child: CircularProgressIndicator(color: Color(0xFF8B4513)),
+              );
+            }
+
+            // 제재된 유저들 (sanctionContent가 있는 경우)
+            final sanctionedDeclarations =
+                controller.declarations
+                    .where(
+                      (d) =>
+                          d.sanctionContent != null &&
+                          d.sanctionContent!.isNotEmpty,
+                    )
+                    .toList();
+
+            if (sanctionedDeclarations.isEmpty) {
+              return Center(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Icon(
+                      Icons.person_off_outlined,
+                      size: 80,
+                      color: Colors.grey[400],
+                    ),
+                    SizedBox(height: 16),
+                    Text(
+                      '제재중인 유저가 없습니다.',
+                      style: TextStyle(fontSize: 16, color: Colors.grey[600]),
+                    ),
+                  ],
+                ),
+              );
+            }
+
+            return RefreshIndicator(
+              onRefresh: controller.refreshData,
+              color: Color(0xFF8B4513),
+              child: ListView.builder(
+                padding: EdgeInsets.all(16),
+                itemCount: sanctionedDeclarations.length,
+                itemBuilder: (context, index) {
+                  final sanction = sanctionedDeclarations[index];
+                  return _buildSanctionedUserItem(sanction);
+                },
               ),
             );
-          }
-
-          return RefreshIndicator(
-            onRefresh: controller.refreshData,
-            color: Color(0xFF8B4513),
-            child: ListView.builder(
-              padding: EdgeInsets.all(16),
-              itemCount: sanctionedDeclarations.length,
-              itemBuilder: (context, index) {
-                final sanction = sanctionedDeclarations[index];
-                return _buildSanctionedUserItem(sanction);
-              },
-            ),
-          );
-        }),
-      ),
-    ],
-  );
-}
+          }),
+        ),
+      ],
+    );
+  }
 
   // 제재 유저 리스트 아이템
   Widget _buildSanctionedUserItem(Declaration sanction) {
@@ -770,7 +808,7 @@ Widget _buildSanctionedUsersTab() {
                 ],
               ),
               SizedBox(height: 16),
-              
+
               // 제재 정보
               Container(
                 padding: EdgeInsets.all(12),
@@ -814,10 +852,7 @@ Widget _buildSanctionedUsersTab() {
                     SizedBox(height: 8),
                     Text(
                       '신고 사유: ${sanction.declarationContent}',
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: Colors.grey[700],
-                      ),
+                      style: TextStyle(fontSize: 12, color: Colors.grey[700]),
                       maxLines: 2,
                       overflow: TextOverflow.ellipsis,
                     ),
@@ -825,14 +860,18 @@ Widget _buildSanctionedUsersTab() {
                 ),
               ),
               SizedBox(height: 12),
-              
+
               // 액션 버튼
               Row(
                 mainAxisAlignment: MainAxisAlignment.end,
                 children: [
                   TextButton.icon(
                     onPressed: () => _showReleaseSanctionDialog(sanction),
-                    icon: Icon(Icons.lock_open, size: 16, color: Colors.green[700]),
+                    icon: Icon(
+                      Icons.lock_open,
+                      size: 16,
+                      color: Colors.green[700],
+                    ),
                     label: Text(
                       '제재 해제',
                       style: TextStyle(
@@ -858,122 +897,123 @@ Widget _buildSanctionedUsersTab() {
   }
 
   // 제재 해제 다이얼로그
-void _showReleaseSanctionDialog(Declaration sanction) {
-  Get.dialog(
-    AlertDialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      title: Row(
-        children: [
-          Icon(Icons.lock_open, color: Colors.green[700]),
-          SizedBox(width: 8),
-          Text('제재 해제'),
-        ],
-      ),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            '다음 사용자의 제재를 해제하시겠습니까?',
-            style: TextStyle(fontSize: 16),
-          ),
-          SizedBox(height: 16),
-          Container(
-            padding: EdgeInsets.all(12),
-            decoration: BoxDecoration(
-              color: Colors.grey[100],
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  '사용자: ${sanction.userId}',
-                  style: TextStyle(fontWeight: FontWeight.w600),
-                ),
-                Text('ID: ${sanction.userId}'),
-                SizedBox(height: 8),
-                Text(
-                  '제재 내용: ${sanction.sanctionContent}',
-                  style: TextStyle(color: Colors.red[700]),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Get.back(),      
-          child: Text('취소', style: TextStyle(color: Colors.grey[600])),
+  void _showReleaseSanctionDialog(Declaration sanction) {
+    Get.dialog(
+      AlertDialog(
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+        title: Row(
+          children: [
+            Icon(Icons.lock_open, color: Colors.green[700]),
+            SizedBox(width: 8),
+            Text('제재 해제'),
+          ],
         ),
-        ElevatedButton(
-          onPressed: () async {
-            try {
-              await controller.releaseSanction(sanction.userId);
-              
-              // 로컬 데이터에서도 해당 Declaration의 sanctionContent를 null로 설정
-              final index = controller.declarations.indexWhere(
-                (d) => d.userId == sanction.userId && d.reviewNum == sanction.reviewNum
-              );
-              
-              if (index != -1) {
-                // 기존 Declaration 객체를 복사하여 sanctionContent를 null로 설정
-                final updatedDeclaration = Declaration(
-                  reviewNum: controller.declarations[index].reviewNum,
-                  userId: controller.declarations[index].userId,
-                  declarationDate: controller.declarations[index].declarationDate,
-                  declarationContent: controller.declarations[index].declarationContent,
-                  declarationState: '처리완료', // 상태는 처리완료로 유지
-                  sanctionContent: null, // 제재 내용을 null로 설정
-                  sanctionDate: controller.declarations[index].sanctionDate,
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text('다음 사용자의 제재를 해제하시겠습니까?', style: TextStyle(fontSize: 16)),
+            SizedBox(height: 16),
+            Container(
+              padding: EdgeInsets.all(12),
+              decoration: BoxDecoration(
+                color: Colors.grey[100],
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    '사용자: ${sanction.userId}',
+                    style: TextStyle(fontWeight: FontWeight.w600),
+                  ),
+                  Text('ID: ${sanction.userId}'),
+                  SizedBox(height: 8),
+                  Text(
+                    '제재 내용: ${sanction.sanctionContent}',
+                    style: TextStyle(color: Colors.red[700]),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Get.back(),
+            child: Text('취소', style: TextStyle(color: Colors.grey[600])),
+          ),
+          ElevatedButton(
+            onPressed: () async {
+              try {
+                await controller.releaseSanction(sanction.userId);
+
+                // 로컬 데이터에서도 해당 Declaration의 sanctionContent를 null로 설정
+                final index = controller.declarations.indexWhere(
+                  (d) =>
+                      d.userId == sanction.userId &&
+                      d.reviewNum == sanction.reviewNum,
                 );
-                
-                // 리스트에서 해당 항목 업데이트
-                controller.declarations[index] = updatedDeclaration;
+
+                if (index != -1) {
+                  // 기존 Declaration 객체를 복사하여 sanctionContent를 null로 설정
+                  final updatedDeclaration = Declaration(
+                    reviewNum: controller.declarations[index].reviewNum,
+                    userId: controller.declarations[index].userId,
+                    declarationDate:
+                        controller.declarations[index].declarationDate,
+                    declarationContent:
+                        controller.declarations[index].declarationContent,
+                    declarationState: '처리완료', // 상태는 처리완료로 유지
+                    sanctionContent: null, // 제재 내용을 null로 설정
+                    sanctionDate: controller.declarations[index].sanctionDate,
+                  );
+
+                  // 리스트에서 해당 항목 업데이트
+                  controller.declarations[index] = updatedDeclaration;
+                }
+
+                // 성공 메시지
+                Get.snackbar(
+                  '완료',
+                  '제재가 해제되었습니다.',
+                  backgroundColor: Colors.green,
+                  colorText: Colors.white,
+                  duration: Duration(seconds: 2),
+                );
+
+                Get.back();
+              } catch (e) {
+                // 에러 발생 시에도 다이얼로그 닫기
+                Get.back();
+
+                // 에러 메시지 표시
+                Get.snackbar(
+                  '오류',
+                  '제재 해제 중 오류가 발생했습니다: ${e.toString()}',
+                  backgroundColor: Colors.red,
+                  colorText: Colors.white,
+                  duration: Duration(seconds: 3),
+                );
               }
-              
-              // 성공 메시지
-              Get.snackbar(
-                '완료',
-                '제재가 해제되었습니다.',
-                backgroundColor: Colors.green,
-                colorText: Colors.white,
-                duration: Duration(seconds: 2),
-              );
-              
-              Get.back();
-            } catch (e) {
-              // 에러 발생 시에도 다이얼로그 닫기
-              Get.back();
-              
-              // 에러 메시지 표시
-              Get.snackbar(
-                '오류',
-                '제재 해제 중 오류가 발생했습니다: ${e.toString()}',
-                backgroundColor: Colors.red,
-                colorText: Colors.white,
-                duration: Duration(seconds: 3),
-              );
-            }
-          },
+            },
             style: ElevatedButton.styleFrom(
-            backgroundColor: Colors.green[700],
-            shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+              backgroundColor: Colors.green[700],
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(10),
+              ),
+            ),
+            child: Text('제재 해제', style: TextStyle(color: Colors.white)),
           ),
-          child: Text('제재 해제', style: TextStyle(color: Colors.white)),
-        ),
-      ],
-    ),
-  );
-}
+        ],
+      ),
+    );
+  }
 
   // 날짜 포맷팅
   String _formatDate(DateTime date) {
     return '${date.year}-${date.month.toString().padLeft(2, '0')}-${date.day.toString().padLeft(2, '0')}';
   }
-
-
 
   // 하단 네비게이션
   Widget _buildBottomNavigation() {
@@ -1013,21 +1053,21 @@ void _showReleaseSanctionDialog(Declaration sanction) {
               ),
             ),
           ),
-          Container(
-            width: 1,
-            height: 40,
-            color: Colors.white.withAlpha(25),
-          ),
+          Container(width: 1, height: 40, color: Colors.white.withAlpha(25)),
           Expanded(
             child: InkWell(
               onTap: () {
                 // 문의 관리 페이지로 이동
-                Get.to(()=>InquiryReport());
+                Get.to(() => InquiryReport());
               },
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  Icon(Icons.question_answer, color: Colors.white.withAlpha(25), size: 26),
+                  Icon(
+                    Icons.question_answer,
+                    color: Colors.white.withAlpha(25),
+                    size: 26,
+                  ),
                   SizedBox(height: 4),
                   Text(
                     '문의 관리',
